+++ src/app/(dashboard)/dashboard/chat/[chatId]/interfaces.tsx
@@ -1,0 +1,17 @@
+import { Message } from "@/lib/validations/messages";
+
+export interface DisplayProps {
+    chatInfo: ChatInfo
+    participants: ChatParticipants
+}
+
+interface ChatParticipants {
+    user: User
+    partner: User
+    sessionId: string
+}
+
+interface ChatInfo {
+    chatId: string
+    messages: Message[]
+}
